backend/requirements.txt
fastapi>=0.115.2
uvicorn[standard]==0.24.0
google-genai>=1.52.0  # Updated package name
pillow>=10.2.0,<11.0.0
numpy>=2.0.0
opencv-python>=4.8.1.78
scikit-image>=0.22.0
pydantic>=2.5.0,<3.0.0
python-multipart>=0.0.6
aiofiles>=23.2.1
requests>=2.31.0
easyocr
